Refactor: Components into microservice architecture
The Webclient is now a seperate component that can be disabled and is not needed for the API to work.
The API also only handles calls to it's respective endpoints.

Added a docker-compose file with an nginx configuration which allows you to bypass cors exceptions when developing locally.

Cleaned up a lot of environment code which only cause noise.

Implemented an environment package handling the reading of the ENV_VARS.